project/requirements.txt
mysql-connector-python==8.0.26
protobuf==3.18.1
urllib3 == 1.26.7